+++ 07.bowling_object/game.rb
@@ -1,13 +1,70 @@
 # frozen_string_literal: true
 
+class Shot
+  attr_reader :mark
+
+  def initialize(mark)
+    @mark = mark
+  end
+
+  def score
+    @mark == 'X' ? 10 : @mark.to_i
+  end
+end
+
+class Frame
+  def initialize(first_mark, second_mark = nil, third_mark = nil)
+    @first_mark = Shot.new(first_mark)
+    @second_mark = Shot.new(second_mark)
+    @third_mark = Shot.new(third_mark)
+  end
+
+  def store_frame_to_array
+    frame_without_nil = [@first_mark.mark, @second_mark.mark, @third_mark.mark]
+    frame_without_nil.include?(nil) ? frame_without_nil.compact! : frame_without_nil
+    case frame_without_nil.length
+    when 1
+      [@first_mark.score]
+    when 2
+      [@first_mark.score, @second_mark.score]
+    when 3
+      [@first_mark.score, @second_mark.score, @third_mark.score]
+    end
+  end
+end
+
 class Game
-  def initialize(frames)
-    @frames = frames
+  def initialize(marks)
+    @marks = marks.split(',')
+  end
+
+  def divide_into_frames
+    frames = []
+    10.times { frames << [] }
+    frames_index = 0
+    @marks.each do |mark|
+      if frames_index == 9
+        frames[frames_index] << (mark == 'X' ? 'X' : mark)
+      elsif mark == 'X'
+        frames[frames_index] << 'X'
+        frames_index += 1
+      else
+        frames[frames_index] << mark
+        frames_index += 1 if frames[frames_index].length == 2
+      end
+    end
+    frames
+  end
+
+  def store_frames_to_array
+    divide_into_frames.map do |frame|
+      Frame.new(frame[0], frame[1], frame[2]).store_frame_to_array
+    end
   end
 
   def calculate_total_score
     total = 0
-    @frames.each_with_index do |frame, index|
+    store_frames_to_array.each_with_index do |frame, index|
       total +=
         if index == 9
           frame.sum
@@ -23,14 +80,18 @@ class Game
   end
 
   def calculate_strike(index, frame)
-    if @frames[index + 1][0] == 10 && index + 1 != 9
-      frame[0] + @frames[index + 1][0] + @frames[index + 2][0]
+    if store_frames_to_array[index + 1][0] == 10 && index + 1 != 9
+      frame[0] + store_frames_to_array[index + 1][0] + store_frames_to_array[index + 2][0]
     else
-      frame[0] + @frames[index + 1][0] + @frames[index + 1][1]
+      frame[0] + store_frames_to_array[index + 1][0] + store_frames_to_array[index + 1][1]
     end
   end
 
   def calculate_spare(index, frame)
-    frame.sum + @frames[index + 1][0]
+    frame.sum + store_frames_to_array[index + 1][0]
   end
 end
+
+game_object = Game.new(ARGV[0])
+total_score = game_object.calculate_total_score
+puts total_score
